requirements.txt
flask==2.2.3
requests==2.28.2
gitpython==3.1.31
flask-socketio==5.3.3
python-socketio==5.8.0
python-engineio==4.4.1
websocket-client==1.5.1
eventlet==0.33.3
gunicorn==20.1.0 